add back the header with all definitions and the const helper
